+++ lib/keyboard_actions_config.dart
@@ -16,10 +16,14 @@ class KeyboardActionsConfig {
   /// Color of the background to the Custom keyboard buttons
   final Color keyboardBarColor;
 
+  /// Color of the line separator between keyboard and content
+  final Color keyboardSeparatorColor;
+
   const KeyboardActionsConfig({
     this.keyboardActionsPlatform = KeyboardActionsPlatform.ALL,
     this.nextFocus = true,
     this.actions,
     this.keyboardBarColor,
+    this.keyboardSeparatorColor = Colors.transparent,
   });
 }